Refactor to keep core.py cleaner
