+++ tools/pre_commit/check_pickle_imports.py
@@ -36,6 +36,7 @@ ALLOWED_FILES = {
     "benchmarks/cutlass_benchmarks/w8a8_benchmarks.py",
     "benchmarks/cutlass_benchmarks/sparse_benchmarks.py",
     "vllm/v1/engine/core.py",  # for serializing inputs in DP gather path
+    "vllm/v1/engine/tt_dp_gather.py",  # extracted DP gather serialization path
     # cloudpickle
     "vllm/v1/executor/multiproc_executor.py",
     "vllm/v1/executor/ray_executor.py",
